+++ src/main/java/com/jhklim/investsim/site/upbit/service/CurrentPriceStore.java
@@ -2,19 +2,20 @@ package com.jhklim.investsim.site.upbit.service;
 
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 @Service
 public class CurrentPriceStore {
 
-    private final Map<String, Double> prices = new ConcurrentHashMap<>();
+    private final Map<String, BigDecimal> prices = new ConcurrentHashMap<>();
 
-    public void update(String market, double price) {
+    public void update(String market, BigDecimal price) {
         prices.put(market, price);
     }
 
-    public double get(String market) {
+    public BigDecimal get(String market) {
         if (!prices.containsKey(market)) {
             throw new IllegalStateException("현재가 없음: " + market);
         }
@@ -24,5 +25,4 @@ public class CurrentPriceStore {
     public boolean exists(String market) {
         return prices.containsKey(market);
     }
-
 }
